adds components to folders, moves state and conditional logic for showing expense form into NewExpense.js
