+++ src/App.js
@@ -7,7 +7,6 @@ import NewExpense from './components/NewExpense/NewExpense';
 function App() {
 
   const [expenses, setExpenses] = useState(expenseData);
-  const [showExpenseForm, setShowExpenseForm] = useState(false);
 
   const addExpenseHandler = (expense) => {
     // add new expense to dummy data
@@ -17,19 +16,12 @@ function App() {
     console.log(expenses);
   }
 
-  const showExpFormHandler = () => {
-    !showExpenseForm ? setShowExpenseForm(() => true) : setShowExpenseForm(() => false);
-  }
-
   return (
     <>
-      {!showExpenseForm ?
-        <button onClick={showExpFormHandler}>Add New Expense</button>
-        :
-        <NewExpense onAddExpense={addExpenseHandler} newExpenseHandler={showExpFormHandler} />}
+      <NewExpense onAddExpense={addExpenseHandler} />
       <Expenses expenses={expenses} />
     </>
   );
 }
 
-export default App;
+export default App
