+++ src/components/Expenses/ExpenseItem/ExpenseItem.js
@@ -1,6 +1,6 @@
 import './ExpenseItem.css';
-import ExpenseDate from './ExpenseDate';
-import Card from '../UI/Card';
+import ExpenseDate from '../ExpenseDate/ExpenseDate';
+import Card from '../../UI/Card';
 import { useState } from 'react';
 
 function ExpenseItem(props) {
