+++ src/components/Expenses/ExpenseList/ExpenseList.js
@@ -1,5 +1,5 @@
 import './ExpenseList.css';
-import ExpenseItem from './ExpenseItem';
+import ExpenseItem from '../ExpenseItem/ExpenseItem';
 
 function ExpenseList(props) {
